Rename core block handler directory
