+++ frontend/components/blocks/core/BlockCover/index.js
