+++ frontend/components/blocks/core/BlockImage/index.js
