+++ frontend/components/blocks/core/BlockImageGallery/index.js
