+++ frontend/components/blocks/core/BlockMediaText/BlockMediaText.js
