+++ frontend/components/blocks/core/BlockParagraph/BlockParagraph.js
