+++ frontend/components/blocks/core/BlockParagraph/index.js
