+++ frontend/components/blocks/core/BlockQuote/BlockQuote.js
