+++ frontend/components/blocks/core/BlockQuote/index.js
