+++ frontend/components/blocks/core/BlockTable/index.js
