+++ frontend/components/blocks/index.js
@@ -1,20 +1,20 @@
 // Export the blocks.
-export {default as BlockButton} from './Gutenberg/BlockButton'
-export {default as BlockButtons} from './Gutenberg/BlockButtons'
-export {default as BlockCode} from './Gutenberg/BlockCode'
-export {default as BlockColumns} from './Gutenberg/BlockColumns'
-export {default as BlockCover} from './Gutenberg/BlockCover'
-export {default as BlockEmbed} from './Gutenberg/BlockEmbed'
-export {default as BlockGravityForm} from './Gutenberg/BlockGravityForm'
-export {default as BlockHeadings} from './Gutenberg/BlockHeadings'
-export {default as BlockImage} from './Gutenberg/BlockImage'
-export {default as BlockImageGallery} from './Gutenberg/BlockImageGallery'
-export {default as BlockList} from './Gutenberg/BlockList'
-export {default as BlockMediaText} from './Gutenberg/BlockMediaText'
-export {default as BlockParagraph} from './Gutenberg/BlockParagraph'
-export {default as BlockPullQuote} from './Gutenberg/BlockPullQuote'
-export {default as BlockQuote} from './Gutenberg/BlockQuote'
-export {default as BlockSeparator} from './Gutenberg/BlockSeparator'
-export {default as BlockShortcode} from './Gutenberg/BlockShortcode'
-export {default as BlockSpacer} from './Gutenberg/BlockSpacer'
-export {default as BlockTable} from './Gutenberg/BlockTable'
+export {default as BlockButton} from './core/BlockButton'
+export {default as BlockButtons} from './core/BlockButtons'
+export {default as BlockCode} from './core/BlockCode'
+export {default as BlockColumns} from './core/BlockColumns'
+export {default as BlockCover} from './core/BlockCover'
+export {default as BlockEmbed} from './core/BlockEmbed'
+export {default as BlockGravityForm} from './core/BlockGravityForm'
+export {default as BlockHeadings} from './core/BlockHeadings'
+export {default as BlockImage} from './core/BlockImage'
+export {default as BlockImageGallery} from './core/BlockImageGallery'
+export {default as BlockList} from './core/BlockList'
+export {default as BlockMediaText} from './core/BlockMediaText'
+export {default as BlockParagraph} from './core/BlockParagraph'
+export {default as BlockPullQuote} from './core/BlockPullQuote'
+export {default as BlockQuote} from './core/BlockQuote'
+export {default as BlockSeparator} from './core/BlockSeparator'
+export {default as BlockShortcode} from './core/BlockShortcode'
+export {default as BlockSpacer} from './core/BlockSpacer'
+export {default as BlockTable} from './core/BlockTable'
